% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computeEventDependentObservation}
\alias{computeEventDependentObservation}
\title{Compute p-value for event-dependent observation end}
\usage{
computeEventDependentObservation(sccsModel, alpha = 0.05)
}
\arguments{
\item{sccsModel}{A fitted SCCS model as created using \code{\link[=fitSccsModel]{fitSccsModel()}}.}

\item{alpha}{The alpha (type 1 error) used to test for exposure rate change.}
}
\value{
A tibble with one row and three columns: \code{ratio} indicates the estimates incidence rate ratio for the
probe at the end of observation. \code{p} is the p-value against the null-hypothesis that the log ratio is
between the \code{endOfObservationEffectBounds} specified when calling \code{fitSccsModel()}, and \code{stable} is
\code{TRUE} if \code{p} is greater than \code{alpha}.
}
\description{
Compute p-value for event-dependent observation end
}
