% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computeEventDependentObservationP}
\alias{computeEventDependentObservationP}
\title{Compute p-value for event-dependent observation end}
\usage{
computeEventDependentObservationP(sccsModel)
}
\arguments{
\item{sccsModel}{A fitted SCCS model as created using \code{\link[=fitSccsModel]{fitSccsModel()}}.}
}
\value{
The p-value
}
\description{
Compute p-value for event-dependent observation end
}
