% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computeExposureChange}
\alias{computeExposureChange}
\title{Compute diagnostic whether exposure probability changed following the outcome}
\usage{
computeExposureChange(
  sccsData,
  studyPopulation,
  exposureEraId = NULL,
  bounds = log(c(1/1.25, 1.25)),
  alpha = 0.05,
  ignoreExposureStarts = FALSE
)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{exposureEraId}{The exposure to create the era data for. If not specified it is
assumed to be the one exposure for which the data was loaded from
the database.}

\item{bounds}{Bounds for the null of no change in the exposure rate.}

\item{alpha}{The alpha (type 1 error) used to test for exposure rate change.}

\item{ignoreExposureStarts}{Ignore exposure starts when computing the diagnostic. This makes the
diagnostic robust against the outcome temporarily preventing exposure
starting, which should be dealt with by the pre-exposure window.}
}
\value{
A tibble with one row and three columns: \code{ratio} indicates the ratio (rate of exposure days after the outcome) /
(rate of exposure days before the outcome). \code{p} is the p-value against the null-hypothesis that the log ratio is
between the provided \code{bounds}, and \code{stable} is \code{TRUE} if \code{p} is greater than \code{alpha}.
}
\description{
Compute diagnostic whether exposure probability changed following the outcome
}
