% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computeExposureChangeP}
\alias{computeExposureChangeP}
\title{Compute p for whether exposure probability changed following the outcome}
\usage{
computeExposureChangeP(
  sccsData,
  studyPopulation,
  exposureEraId = NULL,
  bounds = log(c(0.5, 2)),
  ignoreExposureStarts = TRUE
)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{exposureEraId}{The exposure to create the era data for. If not specified it is
assumed to be the one exposure for which the data was loaded from
the database.}

\item{bounds}{Bounds for the null of no change in the exposure rate.}

\item{ignoreExposureStarts}{Ignore exposure starts when computing the diagnostic. This makes the
diagnostic robust against the outcome temporarily preventing exposure
starting, which should be dealt with by the pre-exposure window.}
}
\value{
The p-value
}
\description{
Compute p for whether exposure probability changed following the outcome
}
