% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computePreExposureGain}
\alias{computePreExposureGain}
\title{Compute pre-exposure risk gain diagnostic}
\usage{
computePreExposureGain(
  sccsData,
  studyPopulation,
  exposureEraId = NULL,
  alpha = 0.05
)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{exposureEraId}{The exposure to create the era data for. If not specified it is
assumed to be the one exposure for which the data was loaded from
the database.}

\item{alpha}{The alpha (type 1 error) used to test for pre-exposure gain}
}
\value{
A tibble with one row and three columns: \code{ratio} indicates the ratio (rate of outcomes before the first exposure start) /
(rate of outcomes after the first exposure start). \code{p} is the p-value against the null-hypothesis that the ratio is
smaller than or equal to 1, and \code{stable} is \code{TRUE} if \code{p} is greater than \code{alpha}.
}
\description{
Compute pre-exposure risk gain diagnostic
}
\details{
Compares the rate of the outcome in the 30 days prior to exposure to the rate
of the outcome in the 30 days following exposure. If the rate before exposure
is higher, this indicates there might reverse causality, that the outcome, or
some precursor of the outcome, increases the probability of having the exposure.

The resulting p-value is computed using a Poisson model conditioned on the person.
}
