% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{computePreExposureGainP}
\alias{computePreExposureGainP}
\title{Compute P for pre-exposure risk gain}
\usage{
computePreExposureGainP(sccsData, studyPopulation, exposureEraId = NULL)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{exposureEraId}{The exposure to create the era data for. If not specified it is
assumed to be the one exposure for which the data was loaded from
the database.}
}
\value{
A one-sided p-value for whether the rate before expore is higher than after, against
the null of no change.
}
\description{
Compute P for pre-exposure risk gain
}
\details{
Compares the rate of the outcome in the 30 days prior to exposure to the rate
of the outcome in the 30 days following exposure. If the rate before exposure
is higher, this indicates there might reverse causality, that the outcome, or
some precursor of the outcome, increases the probability of having the exposure.

The resulting p-value is computed using a Poisson model conditioned on the person.
}
