% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{convertJsonToSccsAnalysesSpecifications}
\alias{convertJsonToSccsAnalysesSpecifications}
\title{Convert JSON to SccsAnalysesSpecifications}
\usage{
convertJsonToSccsAnalysesSpecifications(json)
}
\arguments{
\item{json}{A string containing the JSON representation.}
}
\value{
An object of type \code{SccsAnalysesSpecifications}.
}
\description{
Convert JSON to SccsAnalysesSpecifications
}
