% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{convertSccsAnalysesSpecificationsToJson}
\alias{convertSccsAnalysesSpecificationsToJson}
\title{Convert SccsAnalysesSpecifications to JSON}
\usage{
convertSccsAnalysesSpecificationsToJson(sccsAnalysesSpecifications)
}
\arguments{
\item{sccsAnalysesSpecifications}{An object of type \code{SccsAnalysesSpecifications} as created by the \code{\link[=createSccsAnalysesSpecifications]{createSccsAnalysesSpecifications()}} function.}
}
\value{
A string containing the JSON representation.
}
\description{
Convert SccsAnalysesSpecifications to JSON
}
