% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convertToCcDInputFile}
\alias{convertToCcDInputFile}
\title{Convert CSV to a CCD input file}
\usage{
convertToCcDInputFile(eraFile, observationPeriodFile, targetFile,
  naivePeriod = 180, covariateWindowStart = 0, covariateWindowEnd = 180,
  firstOutcomeOnly = TRUE, minCovariateSubjects = 100)
}
\arguments{
\item{eraFile}{The full path to the era file}

\item{observationPeriodFile}{The full path to the observation period file}

\item{targetFile}{The full path where the CCD input file will be created}

\item{naivePeriod}{Number of days at the start of the observation period that
will not be included in the Poisson regression, but will be used to determine whether an
outcome is a first outcome}

\item{covariateWindowStart}{Number of days relative to covariate start when the risk window
starts. For example, if 0 then the first day of exposure to a drug will be included in the risk
window.}

\item{covariateWindowEnd}{Number of days relative to covariate end when the risk window ends}

\item{firstOutcomeOnly}{Should we include only the first occurrence of the outcome, or all?}

\item{minCovariateSubjects}{Any covariate that is non-zero for less than this number of
subject will be removed from the dataset.}
}
\description{
\code{convertToCcDInputFile} converts era and observation period CSV file to a CCD input file.
}

