% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateSccsEraDataArgs}
\alias{createCreateSccsEraDataArgs}
\title{Create a parameter object for the function createSccsEraData}
\usage{
createCreateSccsEraDataArgs(naivePeriod = 0, firstOutcomeOnly = FALSE,
  covariateSettings, ageSettings = createAgeSettings(includeAge = FALSE),
  seasonalitySettings = createSeasonalitySettings(includeSeasonality = FALSE),
  minCasesForAgeSeason = 10000, eventDependentObservation = FALSE)
}
\arguments{
\item{naivePeriod}{The number of days at the start of a patient's observation
periodthat should not be included in the risk calculations. Note
thatthe naive period can be used to determine current
covariatestatus right after the naive period, and whether an
outcome isthe first one.}

\item{firstOutcomeOnly}{Whether only the first occurrence of an outcome should
beconsidered.}

\item{covariateSettings}{Either an object of type covariateSettings as createdusing the
createCovariateSettings function, or alist of such objects.}

\item{ageSettings}{An object of type ageSettings as created using
thecreateAgeSettings function.}

\item{seasonalitySettings}{An object of type seasonalitySettings as created using
thecreateSeasonalitySettings function.}

\item{minCasesForAgeSeason}{Minimum number of cases to use to fit age and season splines.
IFneeded (and available), cases that are not exposed will be
included.#'}

\item{eventDependentObservation}{Should the extension proposed by Farrington et al. be used
toadjust for event-dependent observation time?}
}
\description{
Create a parameter object for the function createSccsEraData
}
\details{
Create an object defining the parameter values.
}
