% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateScriIntervalDataArgs}
\alias{createCreateScriIntervalDataArgs}
\title{Create a parameter object for the function createScriIntervalData}
\usage{
createCreateScriIntervalDataArgs(eraCovariateSettings, controlIntervalSettings)
}
\arguments{
\item{eraCovariateSettings}{Either an object of type EraCovariateSettings as created using the createEraCovariateSettings() function, or a list of such objects.}

\item{controlIntervalSettings}{An object of type ControlIntervalSettings as created using the createControlIntervalSettings() function.}
}
\description{
Create a parameter object for the function createScriIntervalData
}
\details{
Create an object defining the parameter values.
}
