% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{createDefaultSccsMultiThreadingSettings}
\alias{createDefaultSccsMultiThreadingSettings}
\title{Create default SelfControlledCaseSeries multi-threading settings}
\usage{
createDefaultSccsMultiThreadingSettings(maxCores)
}
\arguments{
\item{maxCores}{Maximum number of CPU cores to use.}
}
\value{
An object of type \code{SccsMultiThreadingSettings}.
}
\description{
Create SelfControlledCaseSeries multi-threading settings based on the maximum
number of cores to be used.
}
\examples{
settings <- createDefaultSccsMultiThreadingSettings(10)

}
\seealso{
\code{\link[=createSccsMultiThreadingSettings]{createSccsMultiThreadingSettings()}}
}
