% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createExposure}
\alias{createExposure}
\title{Create exposure definition}
\usage{
createExposure(exposureId, trueEffectSize = NA)
}
\arguments{
\item{exposureId}{An integer used to identify the exposure in the exposure cohort table.}

\item{trueEffectSize}{For negative and positive controls: the known true effect size. To be
used for empirical calibration. Negative controls have
\code{trueEffectSize = 1}. If the true effect size is unknown, use
\code{trueEffectSize = NA}.}
}
\value{
An object of type \code{exposure}.
}
\description{
Create exposure definition
}
\details{
Create an exposur definition, to be used with the \link{createExposuresOutcome} function.
}
