% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createExposureOutcome}
\alias{createExposureOutcome}
\title{Create a exposure-outcome combination.}
\usage{
createExposureOutcome(exposureId, outcomeId, ...)
}
\arguments{
\item{exposureId}{A concept ID indentifying the target drug in the exposure
table. If multiple strategies for picking the exposure will be
tested in the analysis, a named list of numbers can be provided
instead. In the analysis, the name of the number to be used can
be specified using the #' \code{exposureType} parameter in the
\code{\link{createSccsAnalysis}} function.}

\item{outcomeId}{A concept ID indentifying the outcome in the outcome table.}

\item{...}{Custom variables, to be used in the analyses.}
}
\description{
Create a exposure-outcome combination.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{\link{runSccsAnalyses}} function.
}

