% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createExposuresOutcome}
\alias{createExposuresOutcome}
\title{Create a exposures-outcome combination.}
\usage{
createExposuresOutcome(outcomeId, exposures, nestingCohortId = NULL)
}
\arguments{
\item{outcomeId}{An integer used to identify the outcome in the outcome cohort table.}

\item{exposures}{A list of object of type \code{Exposure} as created by \code{createExposure()}.}

\item{nestingCohortId}{(Optional) the nesting cohort ID.}
}
\value{
An object of type \code{ExposuresOutcome}.
}
\description{
Create a exposures-outcome combination.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{runSccsAnalyses()} function.
}
