% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitSccsModelArgs}
\alias{createFitSccsModelArgs}
\title{Create a parameter object for the function fitSccsModel}
\usage{
createFitSccsModelArgs(
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", selectorType = "byPid", startingVariance =
    0.1, seed = 1, noiseLevel = "quiet"),
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10))
)
}
\arguments{
\item{prior}{The prior used to fit the model. See Cyclops::createPrior for details.}

\item{control}{The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See Cyclops::createControl for details.}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where the likelihood for coefficient of variables is sampled. See details.}

\item{profileBounds}{The bounds (on the log relative risk scale) for the adaptive sampling of the likelihood function.}
}
\description{
Create a parameter object for the function fitSccsModel
}
\details{
Create an object defining the parameter values.
}
