% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbSccsDataArgs}
\alias{createGetDbSccsDataArgs}
\title{Create a parameter object for the function getDbSccsData}
\usage{
createGetDbSccsDataArgs(
  useCustomCovariates = FALSE,
  useNestingCohort = FALSE,
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 100,
  studyStartDate = "",
  studyEndDate = "",
  maxCasesPerOutcome = 0,
  exposureIds = "exposureId",
  customCovariateIds = ""
)
}
\arguments{
\item{useCustomCovariates}{Create covariates from a custom table?}

\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people witha
specific indication)? If not, the study will be nestedin the
general population.}

\item{nestingCohortId}{A cohort definition ID identifying the records in
thenestingCohortTable to use as nesting cohort.}

\item{deleteCovariatesSmallCount}{The minimum count for a covariate to appear in the data to
bekept.}

\item{studyStartDate}{A calendar date specifying the minimum date where data isused.
Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date where data isused.
Date format is 'yyyymmdd'.}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a singleoutcome
cases will be sampled to this size. maxCasesPerOutcome =
0indicates no maximum size.}

\item{exposureIds}{A list of identifiers to define the exposures of interest.
IfexposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID.If
exposureTable <> DRUG_ERA, exposureIds is used to selectthe
cohort_concept_id in the cohort-like table. If noexposureIds are
provided, all drugs or cohorts in theexposureTable are included
as exposures.}

\item{customCovariateIds}{A list of cohort definition IDS identifying the records inthe
customCovariateTable to use for building customcovariates.}
}
\description{
Create a parameter object for the function getDbSccsData
}
\details{
Create an object defining the parameter values.
}
