% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{createSccsDiagnosticThresholds}
\alias{createSccsDiagnosticThresholds}
\title{Create SCCS diagnostics thresholds}
\usage{
createSccsDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  timeTrendPThreshold = 0.05,
  preExposurePThreshold = 0.05
)
}
\arguments{
\item{mdrrThreshold}{What is the maximum allowed minimum detectable relative risk
(MDRR)?}

\item{easeThreshold}{What is the maximum allowed expected absolute systematic error
(EASE).}

\item{timeTrendPThreshold}{What family-wise p-value threshold (alpha) will be used to determine
temporal instability?}

\item{preExposurePThreshold}{What p-value threshold (alpha) will be used to determine whether the
rate of the outcome was higher just before exposure initiation?}
}
\value{
An object of type \code{SccsDiagnosticThresholds}.
}
\description{
Threshold used when calling \code{\link[=exportToCsv]{exportToCsv()}} to determine if we pass or fail diagnostics.
}
