% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsDataConversion.R
\name{createSccsIntervalData}
\alias{createSccsIntervalData}
\title{Create SCCS era data}
\usage{
createSccsIntervalData(studyPopulation, sccsData, createSccsIntervalDataArgs)
}
\arguments{
\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{createSccsIntervalDataArgs}{An object of type \code{CreateSccsIntervalDataArgs} as created by the
\code{createCreateSccsIntervalDataArgs} function.}
}
\value{
An object of type \code{SccsIntervalData}.
}
\description{
Create SCCS era data
}
\details{
This function creates covariates based on the data in the \code{sccsData} argument, according to the
provided settings. It chops patient time into periods during which all covariates remain constant.
The output details these periods, their durations, and a sparse representation of the covariate
values.
}
