% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{createSccsMultiThreadingSettings}
\alias{createSccsMultiThreadingSettings}
\title{Create SelfControlledCaseSeries multi-threading settings}
\usage{
createSccsMultiThreadingSettings(
  getDbSccsDataThreads = 1,
  createStudyPopulationThreads = 1,
  createIntervalDataThreads = 1,
  fitSccsModelThreads = 1,
  cvThreads = 1,
  calibrationThreads = 1
)
}
\arguments{
\item{getDbSccsDataThreads}{The number of parallel threads to use for building the
\code{SccsData} objects.}

\item{createStudyPopulationThreads}{The number of parallel threads to use for building the
\code{studyPopulation} objects.}

\item{createIntervalDataThreads}{The number of parallel threads to use for building the
\code{SccsIntervalData} objects.}

\item{fitSccsModelThreads}{The number of parallel threads to use for fitting the
models.}

\item{cvThreads}{The number of parallel threads to use for the cross-
validation when estimating the hyperparameter for the
outcome model. Note that the total number of CV threads at
one time could be \code{fitSccsModelThreads * cvThreads}.}

\item{calibrationThreads}{The number of parallel threads to use for empirical calibration.}
}
\value{
An object of type \code{SccsMultiThreadingSettings}.
}
\description{
Create SelfControlledCaseSeries multi-threading settings
}
\seealso{
\code{\link[=createDefaultSccsMultiThreadingSettings]{createDefaultSccsMultiThreadingSettings()}}
}
