% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createSimulationRiskWindow}
\alias{createSimulationRiskWindow}
\title{Create a risk window definition for simulation}
\usage{
createSimulationRiskWindow(
  start = 0,
  end = 0,
  endAnchor = "era end",
  splitPoints = c(),
  relativeRisks = c(0)
)
}
\arguments{
\item{start}{Start of the risk window relative to exposure start.}

\item{end}{The end of the risk window (in days) relative to the \code{endAnchor}.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be \code{"era start"}
or \code{"era end"}.}

\item{splitPoints}{Subdivision of the risk window in to smaller sub-windows.}

\item{relativeRisks}{Either a single number representing the relative risk in the risk
window, or when splitPoints have been defined a vector of relative
risks, one for each sub-window.}
}
\value{
An object of type \code{SimulationRiskWindow}.
}
\description{
Create a risk window definition for simulation
}
