% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{createStudyPopulation}
\alias{createStudyPopulation}
\title{Create a study population}
\usage{
createStudyPopulation(sccsData, outcomeId = NULL, createStudyPopulationArgs)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{outcomeId}{The outcome to create the era data for. If not specified it is
assumed to be the one outcome for which the data was loaded from
the database.}

\item{createStudyPopulationArgs}{An object of type \code{CreateStudyPopulationArgs} as created using the
\code{createCreateStudyPopulationArgs()} function.}
}
\value{
A \code{list} specifying the study population, with the following items:
\itemize{
\item \code{cases}: A \code{tibble} with one row per observation period of a person with the outcome.
\item \code{outcomes}: A \code{tibble} listing the days when a case has the outcome.
\item \code{metaData}: A \code{list} with meta data about the study population, including the attrition.
}
}
\description{
Create a study population
}
\details{
Create a study population for a specific outcome, applying several restrictions.
}
