% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsDataConversion.R
\name{cyclicSplineDesign}
\alias{cyclicSplineDesign}
\title{Create a design matrix for a cyclic spline}
\usage{
cyclicSplineDesign(x, knots, ord = 4)
}
\arguments{
\item{x}{Vector of coordinates of the points to be interpolated.}

\item{knots}{Location of the knots.}

\item{ord}{Order of the spline function.}
}
\description{
Create a design matrix for a cyclic spline
}
\details{
This function is used by other functions in this package.
}
