% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R
\name{deleteSccsEraData}
\alias{deleteSccsEraData}
\title{Delete an sccsEraData object from disk.}
\usage{
deleteSccsEraData(sccsEraData, fromCompressedOnly = TRUE)
}
\arguments{
\item{sccsEraData}{The object to delete.}

\item{fromCompressedOnly}{When TRUE, a warning will be thrown when attempting to delete an object that did not
originate from a compressed file, and the delete is canceled.}
}
\description{
Delete an sccsEraData object from disk.
}
\details{
When an object is loaded from a compressed file it is effectively copied. The copy has a tendency to persist.
This function can be used to delete the copy from disk.
}
