% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R
\name{forceSccsEraDataIntoRam}
\alias{forceSccsEraDataIntoRam}
\title{Force a loaded SCCS era data in RAM}
\usage{
forceSccsEraDataIntoRam(sccsEraData)
}
\arguments{
\item{sccsEraData}{Existing sccsEraData object.}
}
\value{
An object of class sccsEraData
}
\description{
\code{forceSccsEraDataIntoRam} converts the ffdf components of an sccsEraData object
into data.table components
}
\details{
Uses ff::as.ram() to move virtual data into data.table objects
}
