% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbSccsData}
\alias{getDbSccsData}
\title{Todo: add title}
\usage{
getDbSccsData(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", outcomeIds,
  outcomeConditionTypeConceptIds = c(),
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  exposureIds = c(), excludeConceptIds = c(), drugEraCovariates = FALSE,
  conditionEraCovariates = FALSE, procedureCovariates = FALSE,
  visitCovariates = FALSE, observationCovariates = FALSE,
  measurementCovariates = FALSE, deleteCovariatesSmallCount = 100,
  cdmVersion = "4")
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{}

\item{oracleTempSchema}{}

\item{outcomeDatabaseSchema}{}

\item{outcomeTable}{}

\item{outcomeIds}{}

\item{outcomeConditionTypeConceptIds}{}

\item{exposureDatabaseSchema}{}

\item{exposureTable}{}

\item{exposureIds}{}

\item{excludeConceptIds}{}

\item{drugEraCovariates}{}

\item{conditionEraCovariates}{}

\item{procedureCovariates}{}

\item{visitCovariates}{}

\item{observationCovariates}{}

\item{measurementCovariates}{}

\item{deleteCovariatesSmallCount}{}

\item{cdmVersion}{}
}
\description{
Todo: add description
}
\details{
Todo: add details
}

