% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{hasCalendarTimeEffect}
\alias{hasCalendarTimeEffect}
\title{Does the model contain an age effect?}
\usage{
hasCalendarTimeEffect(sccsModel)
}
\arguments{
\item{sccsModel}{An object of type \code{SccsModel} as created using the
\code{\link[=fitSccsModel]{fitSccsModel()}} function.}
}
\value{
TRUE if the model contains an age effect, otherwise FALSE.
}
\description{
Does the model contain an age effect?
}
