% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{hasSeasonality}
\alias{hasSeasonality}
\title{Does the model contain an age effect?}
\usage{
hasSeasonality(sccsModel)
}
\arguments{
\item{sccsModel}{An object of type \code{sccsModel} as created using the
\code{\link{fitSccsModel}} function.}
}
\value{
TRUE if the model contains an age effect, otherwise FALSE.
}
\description{
Does the model contain an age effect?
}
