% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\name{isSccsData}
\alias{isSccsData}
\title{Check whether an object is a SccsData object}
\usage{
isSccsData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a SccsData object
}
