% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsEraData.R
\name{isSccsEraData}
\alias{isSccsEraData}
\title{Check whether an object is a SccsEraData object}
\usage{
isSccsEraData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a SccsEraData object
}
