% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsIntervalData.R
\name{isSccsIntervalData}
\alias{isSccsIntervalData}
\title{Check whether an object is a SccsIntervalData object}
\usage{
isSccsIntervalData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a SccsIntervalData object
}
