% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadExposureOutcomeList}
\alias{loadExposureOutcomeList}
\title{Load a list of exposureOutcome from file}
\usage{
loadExposureOutcomeList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{exposureOutcome}.
}
\description{
Load a list of objects of type \code{exposureOutcome} from file. The file is in JSON format.
}

