% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadSccsAnalysisList}
\alias{loadSccsAnalysisList}
\title{Load a list of sccsAnalysis from file}
\usage{
loadSccsAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{SccsAnalysis}.
}
\description{
Load a list of objects of type \code{SccsAnalysis} from file. The file is in JSON format.
}
