% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{loadSccsData}
\alias{loadSccsData}
\title{Load the SCCS data from a folder}
\usage{
loadSccsData(folder, readOnly = FALSE)
}
\arguments{
\item{readOnly}{If true, the data is opened read only.}

\item{file}{The name of the folder containing the data.}
}
\value{
An object of class cohortData.
}
\description{
\code{loadSccsData} loads an object of type sccsData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
#todo
}

