% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsEraData.R
\name{loadSccsEraData}
\alias{loadSccsEraData}
\title{Load the cohort method data from a file}
\usage{
loadSccsEraData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \link{SccsEraData}.
}
\description{
Loads an object of type \link{SccsEraData} from a file in the file system.
}
