% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsIntervalData.R
\name{loadSccsIntervalData}
\alias{loadSccsIntervalData}
\title{Load the cohort method data from a file}
\usage{
loadSccsIntervalData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \link{SccsIntervalData}.
}
\description{
Loads an object of type \link{SccsIntervalData} from a file in the file system.
}
