% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotEventToCalendarTime}
\alias{plotEventToCalendarTime}
\title{Plot the ratio of observed to expected events over calendar time.}
\usage{
plotEventToCalendarTime(
  studyPopulation,
  sccsModel = NULL,
  title = NULL,
  fileName = NULL
)
}
\arguments{
\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{sccsModel}{Optional: A fitted SCCS model as created using \code{\link[=fitSccsModel]{fitSccsModel()}}. If the
model contains splines for seasonality and or calendar time a panel will
be added with outcome counts adjusted for these splines.}

\item{title}{Optional: the main title for the plot}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} for supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function to save to file in a different
format.
}
\description{
Plot the ratio of observed to expected events over calendar time.
}
\details{
Plot the ratio of observed to expected events over calendar time. The expected count expected
rate considers which persons were observed during that month, and if specified in the model,
the adjustment for season and calendar time.
}
