% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotOutcomeCentered}
\alias{plotOutcomeCentered}
\title{Plot exposure days relative to the first outcome}
\usage{
plotOutcomeCentered(
  studyPopulation,
  sccsData,
  exposureEraId = NULL,
  title = NULL,
  fileName = NULL
)
}
\arguments{
\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{exposureEraId}{The exposure to create the era data for. If not specified it is
assumed to be the one exposure for which the data was loaded from
the database.}

\item{title}{Optional: the main title for the plot}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} for supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function to save to file in a different
format.
}
\description{
Plot exposure days relative to the first outcome
}
