% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{saveExposuresOutcomeList}
\alias{saveExposuresOutcomeList}
\title{Save a list of \code{ExposuresOutcome} to file}
\usage{
saveExposuresOutcomeList(exposuresOutcomeList, file)
}
\arguments{
\item{exposuresOutcomeList}{A list of objects of type \code{ExposuresOutcome} as created using the \code{\link[=createExposuresOutcome]{createExposuresOutcome()}} function.}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{ExposuresOutcome} to file. The file is in JSON format.
}
