% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\name{saveSccsData}
\alias{saveSccsData}
\title{Save the cohort method data to file}
\usage{
saveSccsData(SccsData, file)
}
\arguments{
\item{SccsData}{An object of type [SccsData] as generated using
[getDbSccsData()].}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type [SccsData] to a file.
}
