% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsEraData.R
\name{saveSccsEraData}
\alias{saveSccsEraData}
\title{Save the cohort method data to file}
\usage{
saveSccsEraData(SccsEraData, file)
}
\arguments{
\item{SccsEraData}{An object of type \link{SccsEraData} as generated using
\code{\link[=createSccsEraData]{createSccsEraData()}}.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type \link{SccsEraData} to a file.
}
