% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsIntervalData.R
\name{saveSccsIntervalData}
\alias{saveSccsIntervalData}
\title{Save the cohort method data to file}
\usage{
saveSccsIntervalData(SccsIntervalData, file)
}
\arguments{
\item{SccsIntervalData}{An object of type \link{SccsIntervalData} as generated using
\code{\link[=createSccsIntervalData]{createSccsIntervalData()}}.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type \link{SccsIntervalData} to a file.
}
