% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\docType{class}
\name{SccsData-class}
\alias{SccsData-class}
\alias{SccsData}
\alias{show,SccsData-method}
\alias{summary,SccsData-method}
\title{SCCS Data}
\usage{
\S4method{show}{SccsData}(object)

\S4method{summary}{SccsData}(object)
}
\arguments{
\item{object}{An object of type `SccsData`.}
}
\description{
`SccsData` is an S4 class that inherits from [Andromeda][Andromeda::Andromeda]. It contains information on the cases and their covariates.

A `SccsData` is typically created using [getDbSccsData()], can only be saved using
[saveSccsData()], and loaded using [loadSccsData()].
}
