% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsEraData.R
\docType{class}
\name{SccsEraData-class}
\alias{SccsEraData-class}
\alias{SccsEraData}
\alias{show,SccsEraData-method}
\alias{summary,SccsEraData-method}
\title{SCCS Era Data}
\usage{
\S4method{show}{SccsEraData}(object)

\S4method{summary}{SccsEraData}(object)
}
\arguments{
\item{object}{An object of type \code{SccsEraData}.}
}
\description{
SccsEraData` is an S4 class that inherits from \link[Andromeda:Andromeda]{Andromeda}. It contains information on the cases and their covariates.

A \code{SccsEraData} is typically created using \code{\link[=createSccsEraData]{createSccsEraData()}}, can only be saved using
\code{\link[=saveSccsEraData]{saveSccsEraData()}}, and loaded using \code{\link[=loadSccsEraData]{loadSccsEraData()}}.
}
