% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsIntervalData.R
\docType{class}
\name{SccsIntervalData-class}
\alias{SccsIntervalData-class}
\alias{SccsIntervalData}
\alias{show,SccsIntervalData-method}
\alias{summary,SccsIntervalData-method}
\title{SCCS Interval Data}
\usage{
\S4method{show}{SccsIntervalData}(object)

\S4method{summary}{SccsIntervalData}(object)
}
\arguments{
\item{object}{An object of type \code{SccsIntervalData}.}
}
\description{
SccsIntervalData` is an S4 class that inherits from \link[Andromeda:Andromeda]{Andromeda}. It contains
information on the cases and their covariates, divided in non-overlapping time intervals.

A \code{SccsIntervalData} is typically created using \code{\link[=createSccsIntervalData]{createSccsIntervalData()}}, can only be saved using
\code{\link[=saveSccsIntervalData]{saveSccsIntervalData()}}, and loaded using \code{\link[=loadSccsIntervalData]{loadSccsIntervalData()}}.
}
