% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SelfControlledCaseSeries.R
\docType{package}
\name{SelfControlledCaseSeries}
\alias{SelfControlledCaseSeries}
\alias{SelfControlledCaseSeries-package}
\title{SelfControlledCaseSeries}
\description{
SelfControlledCaseSeries
}

