% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateSccsData}
\alias{simulateSccsData}
\title{Simulate SCCS data}
\usage{
simulateSccsData(nCases, settings)
}
\arguments{
\item{nCases}{The number of cases to simulate.}

\item{settings}{An object of type \code{SccsSimulationSettings} as created using the
\code{\link[=createSccsSimulationSettings]{createSccsSimulationSettings()}} function.}
}
\value{
An object of type \code{SccsData}.
}
\description{
Simulate SCCS data
}
