% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizeSccsAnalyses}
\alias{summarizeSccsAnalyses}
\title{Create a summary report of the analyses}
\usage{
summarizeSccsAnalyses(referenceTable, outputFolder)
}
\arguments{
\item{referenceTable}{A tibble as created by the \link{runSccsAnalyses} function.}

\item{outputFolder}{Name of the folder where all the outputs have been written to.}
}
\value{
A tibble containing summary statistics for each exposure-outcome-analysis combination.
}
\description{
Create a summary report of the analyses
}
