% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateSettings.R
\name{createCalendarTimeCovariateSettings}
\alias{createCalendarTimeCovariateSettings}
\title{Create calendar time settings}
\usage{
createCalendarTimeCovariateSettings(
  calendarTimeKnots = 5,
  allowRegularization = FALSE,
  computeConfidenceIntervals = FALSE
)
}
\arguments{
\item{calendarTimeKnots}{If a single number is provided this is assumed to indicate the
number of knots to use for the spline, and the knots are
automatically spaced according to equal percentiles of the data.
If a series of dates is provided these are assumed to be the exact location of
the knots.}

\item{allowRegularization}{When fitting the model, should the covariates defined here be
allowed to be regularized?}

\item{computeConfidenceIntervals}{Should confidence intervals be computed for the covariates
defined here? Setting this to FALSE might save computing time
when fitting the model. Will be turned to FALSE automatically
when \code{allowRegularization = TRUE}.}
}
\value{
An object of type \code{seasonalitySettings}.
}
\description{
Create calendar time settings
}
\details{
Create an object specifying whether and how calendar time should be included in the model.
Calendar time can be included by splitting patient time into calendar months. During a month, the
relative risk attributed to calendar time is assumed to be constant, and the risk from month to month is
modeled using a cubic spline.

Whereas the seasonality covariate uses a cyclic spline, repeating every year, this calendar time covariate
can model trends over years.
}
