% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createExposure}
\alias{createExposure}
\title{Create exposure definition}
\usage{
createExposure(exposureId, exposureIdRef = "exposureId", trueEffectSize = NA)
}
\arguments{
\item{exposureId}{An integer used to identify the exposure in the exposure cohort table.}

\item{exposureIdRef}{A string used to refer to the exposure when defining covariates using the
\code{createEraCovariateSettings()} function.}

\item{trueEffectSize}{For negative and positive controls: the known true effect size. To be
used for empirical calibration. Negative controls have
\code{trueEffectSize = 1}. If the true effect size is unknown, use
\code{trueEffectSize = NA}.}
}
\value{
An object of type \code{Exposure}.
}
\description{
Create exposure definition
}
\details{
Create an exposure definition, to be used with the \link{createExposuresOutcome} function.
}
