% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createExposuresOutcome}
\alias{createExposuresOutcome}
\title{Create a exposures-outcome combination.}
\usage{
createExposuresOutcome(outcomeId, exposures)
}
\arguments{
\item{outcomeId}{An integer used to identify the outcome in the outcome cohort table.}

\item{exposures}{A list of object of type \code{Exposure} as created by \code{\link[=createExposure]{createExposure()}}.}
}
\value{
An object of type \code{ExposuresOutcome}.
}
\description{
Create a exposures-outcome combination.
}
\details{
Create a set of hypotheses of interest, to be used with the \link{runSccsAnalyses} function.
}
