% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbSccsDataArgs}
\alias{createGetDbSccsDataArgs}
\title{Create a parameter object for the function getDbSccsData}
\usage{
createGetDbSccsDataArgs(
  useCustomCovariates = FALSE,
  useNestingCohort = FALSE,
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 0,
  studyStartDate = "",
  studyEndDate = "",
  maxCasesPerOutcome = 0,
  exposureIds = "exposureId",
  customCovariateIds = ""
)
}
\arguments{
\item{useCustomCovariates}{Create covariates from a custom table?}

\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people with a specific indication)? If not, the study will be nested in the general population.}

\item{nestingCohortId}{A cohort definition ID identifying the records in the nestingCohortTable to use as nesting cohort.}

\item{deleteCovariatesSmallCount}{The minimum count for a covariate to appear in the data to be kept.}

\item{studyStartDate}{A character object specifying the minimum date where data is used. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A character object specifying the maximum date where data is used. Date format is 'yyyymmdd'.}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a single outcome cases will be sampled to this size. maxCasesPerOutcome = 0 indicates no maximum size.}

\item{exposureIds}{A list of identifiers to extract from the exposure table. If exposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID. If exposureTable = "drug_era", exposureIds is used to select the drug_concept_id. If no exposure IDs are provided, all drugs or cohorts in the exposureTable are included as exposures.}

\item{customCovariateIds}{A list of cohort definition IDs identifying the records in the customCovariateTable to use for building custom covariates.}
}
\description{
Create a parameter object for the function getDbSccsData
}
\details{
Create an object defining the parameter values.
}
