% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createSccsAnalysis}
\alias{createSccsAnalysis}
\title{Create a SelfControlledCaseSeries analysis specification}
\usage{
createSccsAnalysis(
  analysisId = 1,
  description = "",
  getDbSccsDataArgs,
  createStudyPopulationArgs,
  createIntervalDataArgs = NULL,
  fitSccsModelArgs
)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific set
of analysis choices.}

\item{description}{A short description of the analysis.}

\item{getDbSccsDataArgs}{An object representing the arguments to be used when calling the
\link{getDbSccsData} function.}

\item{createStudyPopulationArgs}{An object representing the arguments to be used when calling the
\link{getDbSccsData} function.}

\item{createIntervalDataArgs}{An object representing the arguments to be used when calling the
\link{createSccsIntervalData} or \link{createScriIntervalData} function.}

\item{fitSccsModelArgs}{An object representing the arguments to be used when calling the
\link{fitSccsModel} function.}
}
\value{
An object of type \code{SccsAnalysis}, to be used with the \link{runSccsAnalyses} function.
}
\description{
Create a SelfControlledCaseSeries analysis specification
}
