% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R
\name{createSccsIntervalData}
\alias{createSccsIntervalData}
\title{Create SCCS era data}
\usage{
createSccsIntervalData(
  studyPopulation,
  sccsData,
  eraCovariateSettings,
  ageCovariateSettings = NULL,
  seasonalityCovariateSettings = NULL,
  minCasesForAgeSeason = 10000,
  eventDependentObservation = FALSE
)
}
\arguments{
\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{eraCovariateSettings}{Either an object of type \code{EraCovariateSettings} as created
using the \code{\link[=createEraCovariateSettings]{createEraCovariateSettings()}} function, or a
list of such objects.}

\item{ageCovariateSettings}{An object of type \code{ageCovariateSettings} as created using the
\code{\link[=createAgeCovariateSettings]{createAgeCovariateSettings()}} function.}

\item{seasonalityCovariateSettings}{An object of type \code{seasonalityCovariateSettings} as created using the
\code{\link[=createSeasonalityCovariateSettings]{createSeasonalityCovariateSettings()}} function.}

\item{minCasesForAgeSeason}{Minimum number of cases to use to fit age and season splines. If
needed (and available), cases that are not exposed will be included.}

\item{eventDependentObservation}{Should the extension proposed by Farrington et al. be used to
adjust for event-dependent observation time?}
}
\value{
An object of type \link{SccsIntervalData}.
}
\description{
Create SCCS era data
}
\details{
This function creates covariates based on the data in the \code{sccsData} argument, according to the
provided settings. It chops patient time into periods during which all covariates remain constant.
The output details these periods, their durations, and a sparse representation of the covariate
values.
}
\references{
Farrington, C. P., Anaya-Izquierdo, A., Whitaker, H. J., Hocine, M.N., Douglas, I., and Smeeth, L.
(2011). Self-Controlled case series analysis with event-dependent observation periods. Journal of
the American Statistical Association 106 (494), 417-426
}
