% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createSccsSimulationSettings}
\alias{createSccsSimulationSettings}
\title{Create SCCS simulation settings}
\usage{
createSccsSimulationSettings(meanPatientTime = 4 * 365,
  sdPatientTime = 2 * 365, minAge = 18 * 365, maxAge = 65 * 365,
  minBaselineRate = 0.001, maxBaselineRate = 0.01,
  covariateIds = c(1, 2), patientUsages = c(0.2, 0.1),
  usageRate = c(0.01, 0.01), meanPrescriptionDurations = c(14, 30),
  sdPrescriptionDurations = c(7, 14),
  simulationRiskWindows = list(createSimulationRiskWindow(relativeRisks =
  1), createSimulationRiskWindow(relativeRisks = 1.5)),
  includeAgeEffect = TRUE, ageKnots = 5, includeSeasonality = TRUE,
  seasonKnots = 5, outcomeId = 10)
}
\arguments{
\item{meanPatientTime}{Mean number of observation days per patient.}

\item{sdPatientTime}{Standard deviation of the observation days per patient.}

\item{minAge}{The minimum age in days.}

\item{maxAge}{The maximum age in days.}

\item{minBaselineRate}{The minimum baseline rate (per day).}

\item{maxBaselineRate}{The maximum baseline rate (per day).}

\item{covariateIds}{The IDs for the covariates to be generated.}

\item{patientUsages}{The fraction of patients that use the drugs.}

\item{usageRate}{The rate of prescriptions per person that uses the drug.}

\item{meanPrescriptionDurations}{The mean duration of a prescription, per drug.}

\item{sdPrescriptionDurations}{The standard deviation of the duration of a prescription, per
drug.}

\item{simulationRiskWindows}{One or a list of objects of type \code{simulationRiskWindow} as
created using the \code{\link{createSimulationRiskWindow}}
function.}

\item{includeAgeEffect}{Include an age effect for the outcome?}

\item{ageKnots}{Number of knots in the age spline.}

\item{includeSeasonality}{Include seasonality for the outcome?}

\item{seasonKnots}{Number of knots in the seasonality spline.}

\item{outcomeId}{The ID to be used for the outcome.}
}
\value{
An object of type \code{sccsSimulationSettings}.
}
\description{
Create SCCS simulation settings
}
\details{
Create an object of settings for an SCCS simulation.
}
