% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbSccsData}
\alias{getDbSccsData}
\title{Load data for SCCS from the database}
\usage{
getDbSccsData(
  connectionDetails,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  outcomeIds,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  exposureIds = c(),
  useCustomCovariates = FALSE,
  customCovariateDatabaseSchema = cdmDatabaseSchema,
  customCovariateTable = "cohort",
  customCovariateIds = c(),
  useNestingCohort = FALSE,
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "cohort",
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 0,
  studyStartDate = "",
  studyEndDate = "",
  cdmVersion = "5",
  maxCasesPerOutcome = 0
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specify both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
\code{outcomeTable = "condition_era"}, \code{outcomeDatabaseSchema} is not
used.  Requires read permissions to this database.}

\item{outcomeTable}{The table name that contains the outcome cohorts.  If
\code{outcomeTable} is not \code{"condition_era"},
then expectation is \code{outcomeTable} has format of cohort table (see
details).}

\item{outcomeIds}{A list of IDs used to define outcomes.  If \code{outcomeTable} is not
\code{"condition_era"} the list contains records found in the
\code{cohort_definition_id} field.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure eras is
available. If \code{exposureTable = "drug_era"},
\code{exposureDatabaseSchema} is not used but assumed to be equal to
\code{cdmDatabaseSchema}.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
\code{exposureTable} is not "drug_era", then expectation is \code{exposureTable}
has format of a cohort table (see details).}

\item{exposureIds}{A list of identifiers to extract from the exposure table. If
exposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID.
If \code{exposureTable = "drug_era"}, \code{exposureIds} is used to select
the \code{drug_concept_id}. If no exposure IDs are provided, all drugs
or cohorts in the \code{exposureTable} are included as exposures.}

\item{useCustomCovariates}{Create covariates from a custom table?}

\item{customCovariateDatabaseSchema}{The name of the database schema that is the location where
the custom covariate data is available.}

\item{customCovariateTable}{Name of the table holding the custom covariates. This table
should have the same structure as the cohort table (see details).}

\item{customCovariateIds}{A list of cohort definition IDs identifying the records in
the \code{customCovariateTable} to use for building custom
covariates.}

\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people with
a specific indication)? If not, the study will be nested
in the general population.}

\item{nestingCohortDatabaseSchema}{The name of the database schema that is the location
where the nesting cohort is defined.}

\item{nestingCohortTable}{Name of the table holding the nesting cohort. This table
should have the same structure as the cohort table (see details).}

\item{nestingCohortId}{A cohort definition ID identifying the records in the
\code{nestingCohortTable} to use as nesting cohort.}

\item{deleteCovariatesSmallCount}{The minimum count for a covariate to appear in the data to be
kept.}

\item{studyStartDate}{A character object specifying the minimum date where data is
used. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A character object specifying the maximum date where data is
used. Date format is 'yyyymmdd'.}

\item{cdmVersion}{Define the OMOP CDM version used: currently supports "5".}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a single
outcome cases will be sampled to this size. \code{maxCasesPerOutcome = 0}
indicates no maximum size.}
}
\value{
An \link{SccsData} object.
}
\description{
Load all data needed to perform an SCCS analysis from the database.
}
\details{
This function downloads several types of information:
\itemize{
\item Information on the occurrences of the outcome(s) of interest. Note that information for
multiple outcomes can be fetched in one go, and later the specific outcome can be specified
for which we want to build a model.
\item Information on the observation time and age for the people with the outcomes.
\item Information on exposures of interest which we want to include in the model.
}

Five different database schemas can be specified, for five different types of information: The
\itemize{
\item \strong{cdmDatabaseSchema} is used to extract patient age and observation period. The
\item \strong{outcomeDatabaseSchema} is used to extract information about the outcomes, the
\item \strong{exposureDatabaseSchema} is used to retrieve information on exposures, and the
\item \strong{customCovariateDatabaseSchema} is optionally used to find additional, user-defined
covariates. All four locations could point to the same database schema.
\item \strong{nestingCohortDatabaseSchema} is optionally used to define a cohort in which the analysis is nested,
for example a cohort of diabetics patients.
}

All five locations could point to the same database schema.

Cohort tables are assumed to have the following fields: \code{cohort_definition_id}, \code{subject_id},
\code{cohort_start_date}, and \code{cohort_end_date.}
}
