% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbSccsData}
\alias{getDbSccsData}
\title{Load data for SCCS from the database}
\usage{
getDbSccsData(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  outcomeIds,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  exposureIds = c(),
  useCustomCovariates = FALSE,
  customCovariateDatabaseSchema = cdmDatabaseSchema,
  customCovariateTable = "cohort",
  customCovariateIds = c(),
  useNestingCohort = FALSE,
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "cohort",
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 100,
  studyStartDate = "",
  studyEndDate = "",
  cdmVersion = "5",
  maxCasesPerOutcome = 0
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specify both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{DEPRECATED: use \code{tempEmulationSchema} instead.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is not
used.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA,
then expectation is outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{outcomeIds}{A list of ids used to define outcomes.  If outcomeTable =
CONDITION_OCCURRENCE, the list is a set of ancestor
CONCEPT_IDs, and all occurrences of all descendant concepts
will be selected.  If outcomeTable <> CONDITION_OCCURRENCE,
the list contains records found in COHORT_DEFINITION_ID
field.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used but assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable
has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{exposureIds}{A list of identifiers to define the exposures of interest. If
exposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID.
If exposureTable <> DRUG_ERA, exposureIds is used to select
the cohort_concept_id in the cohort-like table. If no
exposureIds are provided, all drugs or cohorts in the
exposureTable are included as exposures.}

\item{useCustomCovariates}{Create covariates from a custom table?}

\item{customCovariateDatabaseSchema}{The name of the database schema that is the location where
the custom covariate data is available.}

\item{customCovariateTable}{Name of the table holding the custom covariates. This table
should have the same structure as the cohort table.}

\item{customCovariateIds}{A list of cohort definition IDS identifying the records in
the customCovariateTable to use for building custom
covariates.}

\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people with
a specific indication)? If not, the study will be nested
in the general population.}

\item{nestingCohortDatabaseSchema}{The name of the database schema that is the location
where the nesting cohort is defined.}

\item{nestingCohortTable}{Name of the table holding the nesting cohort. This table
should have the same structure as the cohort table.}

\item{nestingCohortId}{A cohort definition ID identifying the records in the
nestingCohortTable to use as nesting cohort.}

\item{deleteCovariatesSmallCount}{The minimum count for a covariate to appear in the data to be
kept.}

\item{studyStartDate}{A calendar date specifying the minimum date where data is
used. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date where data is
used. Date format is 'yyyymmdd'.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a single
outcome cases will be sampled to this size. \code{maxCasesPerOutcome = 0}
indicates no maximum size.}
}
\value{
An \link{SccsData} object.
}
\description{
Load all data needed to perform an SCCS analysis from the database.
}
\details{
This function downloads several types of information:
\itemize{
\item Information on the occurrences of the outcome(s) of interest. Note that information for
multiple outcomes can be fetched in one go, and later the specific outcome can be specified
for which we want to build a model.
\item Information on the observation time and age for the people with the outcomes.
\item Information on exposures of interest which we want to include in the model.
}

Five different database schemas can be specified, for five different types of information: The
\itemize{
\item \strong{cdmDatabaseSchema} is used to extract patient age and observation period. The
\item \strong{outcomeDatabaseSchema} is used to extract information about the outcomes, the
\item \strong{exposureDatabaseSchema} is used to retrieve information on exposures, and the
\item \strong{customCovariateDatabaseSchema} is optionally used to find additional, user-defined
covariates. All four locations could point to the same database schema.
\item \strong{nestingCohortDatabaseSchema} is optionally used to define a cohort in which the analysis is nested,
for example a cohort of diabetics patients.
}

All five locations could point to the same database schema.
}
