% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbSccsData}
\alias{getDbSccsData}
\title{Todo: add title}
\usage{
getDbSccsData(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", outcomeConceptIds,
  outcomeConditionTypeConceptIds = c(),
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  exposureConceptIds = c(), excludeConceptIds = c(),
  drugEraCovariates = FALSE, conditionEraCovariates = FALSE,
  procedureCovariates = FALSE, visitCovariates = FALSE,
  observationCovariates = FALSE, deleteCovariatesSmallCount = 100)
}
\arguments{
\item{cdmDatabaseSchema}{}

\item{oracleTempSchema}{}

\item{outcomeDatabaseSchema}{}

\item{outcomeTable}{}

\item{outcomeConceptIds}{}

\item{outcomeConditionTypeConceptIds}{}

\item{exposureDatabaseSchema}{}

\item{exposureTable}{}

\item{exposureConceptIds}{}

\item{excludeConceptIds}{}

\item{drugEraCovariates}{}

\item{conditionEraCovariates}{}

\item{procedureCovariates}{}

\item{visitCovariates}{}

\item{observationCovariates}{}

\item{deleteCovariatesSmallCount}{}

\item{connectionDetailsAn}{R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}
}
\description{
Todo: add description
}
\details{
Todo: add details
}

