% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbSccsData}
\alias{getDbSccsData}
\title{Todo: add title}
\usage{
getDbSccsData(connectionDetails, cdmDatabaseSchema = "cdm4_sim.dbo",
  resultsDatabaseSchema = "scratch.dbo",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", outcomeConceptIds = "",
  outcomeConditionTypeConceptIds = "",
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  exposureConceptIds = "", drugEraCovariates = FALSE,
  conditionEraCovariates = FALSE, procedureCovariates = FALSE,
  visitCovariates = FALSE, observationCovariates = FALSE,
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{}

\item{resultsDatabaseSchema}{}

\item{outcomeDatabaseSchema}{}

\item{outcomeTable}{}

\item{outcomeConceptIds}{}

\item{outcomeConditionTypeConceptIds}{}

\item{exposureDatabaseSchema}{}

\item{exposureTable}{}

\item{exposureConceptIds}{}

\item{drugEraCovariates}{}

\item{conditionEraCovariates}{}

\item{procedureCovariates}{}

\item{visitCovariates}{}

\item{observationCovariates}{}
}
\description{
Todo: add description
}
\details{
Todo: add details
}

