% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{getModel}
\alias{getModel}
\title{Output the full model}
\usage{
getModel(sccsModel)
}
\arguments{
\item{sccsModel}{An object of type \code{SccsModel} as created using the
\code{\link[=fitSccsModel]{fitSccsModel()}} function.}
}
\value{
A \code{tibble} with the coefficients and confidence intervals (when not-regularized) for all
covariates in the model.
}
\description{
Output the full model
}
