% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadExposuresOutcomeList}
\alias{loadExposuresOutcomeList}
\title{Load a list of \code{exposuresOutcome} from file}
\usage{
loadExposuresOutcomeList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{exposuresOutcome}.
}
\description{
Load a list of objects of type \code{exposuresOutcome} from file. The file is in JSON format.
}
