% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\name{loadSccsData}
\alias{loadSccsData}
\title{Load the cohort method data from a file}
\usage{
loadSccsData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \link{SccsData}.
}
\description{
Loads an object of type \link{SccsData} from a file in the file system.
}
