% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R, R/SccsEraData.R
\name{loadSccsEraData}
\alias{loadSccsEraData}
\title{Load the SCCS era data from a folder}
\usage{
loadSccsEraData(file)

loadSccsEraData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}

\item{folder}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class sccsEraData

An object of class \link{SccsEraData}.
}
\description{
\code{loadSccsEraData} loads an object of type sccsEraData from a folder in the file system.

Loads an object of type \link{SccsEraData} from a file in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
