% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R
\name{loadSccsEraData}
\alias{loadSccsEraData}
\title{Load the SCCS era data from a folder}
\usage{
loadSccsEraData(folder, readOnly = FALSE)
}
\arguments{
\item{folder}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class sccsEraData
}
\description{
\code{loadSccsEraData} loads an object of type sccsEraData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}

