% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotEventObservationDependence}
\alias{plotEventObservationDependence}
\title{Plot time from event to observation end for censored and uncensored time.}
\usage{
plotEventObservationDependence(
  sccsData,
  outcomeId = NULL,
  naivePeriod = 0,
  minAge = NULL,
  maxAge = NULL,
  fileName = NULL
)
}
\arguments{
\item{sccsData}{An object of type \code{sccsData} as created using the
\code{\link{getDbSccsData}} function.}

\item{outcomeId}{The outcome to create the era data for. If not specified it is
assumed to be the one outcome for which the data was loaded from
the database.}

\item{naivePeriod}{The number of days at the start of a patient's observation period
that should not be included in the risk calculations. Note that
the naive period can be used to determine current covariate
status right after the naive period, and whether an outcome is
the first one.}

\item{minAge}{Minimum age at which patient time will be included in the analysis. Note
that information prior to the min age is still used to determine exposure
status after the minimum age (e.g. when a prescription was started just prior
to reaching the minimum age). Also, outcomes occurring before the minimum age
is reached will be considered as prior outcomes when using first outcomes only.
Age should be specified in years, but non-integer values are allowed. If not
specified, no age restriction will be applied.}

\item{maxAge}{Maximum age at which patient time will be included in the analysis. Age should
be specified in years, but non-integer values are allowed. If not
specified, no age restriction will be applied.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A Ggplot object. Use the ggsave function to save to file.
}
\description{
Plot time from event to observation end for censored and uncensored time.
}
\details{
This plot shows whether there is a difference in time between (first) event and the observation period end for periods that are '
censored' and those that are 'uncensored'. By 'censored' we mean periods that end before we would normally expect. Here, we define
periods to be uncensored if they end at either the study end date (if specified), database end date (i.e. the date after which no
data is captured in the database), or maximum age (if specified). All other periods are assumed to be censored.

As proposed by Farrington et al., by comparing the two plots, we can gain some insight into whether the censoring is dependent
on the occurrence of the event.

If parameters such as naivePeriod, minAge, and maxAge are provided, these will first be applied to curtail the observation period
prior to plotting. Similarly, firstOutcomeOnly can be provided so subjects where the first outcome falls before the true observation
start are removed before plotting.
}
\references{
Farrington P, Whitaker H, Ghebremichael Weldeselassie Y (2018), Self-controlled case series studies: A modelling guide with R, Taylor &
Francis
}
