% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotEventToCalendarTime}
\alias{plotEventToCalendarTime}
\title{Plot the count of events over calendar time.}
\usage{
plotEventToCalendarTime(studyPopulation, fileName = NULL)
}
\arguments{
\item{studyPopulation}{An object created using the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} for supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function to save to file in a different
format.
}
\description{
Plot the count of events over calendar time.
}
