% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runSccsAnalyses}
\alias{runSccsAnalyses}
\title{Run a list of analyses}
\usage{
runSccsAnalyses(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema, outcomeTable = "condition_era",
  customCovariateDatabaseSchema = cdmDatabaseSchema,
  customCovariateTable = "cohort", cdmVersion = 5,
  outputFolder = "./SccsOutput", sccsAnalysisList, exposureOutcomeList,
  combineDataFetchAcrossOutcomes = TRUE, getDbSccsDataThreads = 1,
  createSccsEraDataThreads = 1, fitSccsModelThreads = 1, cvThreads = 1)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used but assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable
has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is not
used.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA,
then expectation is outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{customCovariateDatabaseSchema}{The name of the database schema that is the location where
the custom covariate data is available.}

\item{customCovariateTable}{Name of the table holding the custom covariates. This table
should have the same structure as the cohort table.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{outputFolder}{Name of the folder where all the outputs will written
to.}

\item{sccsAnalysisList}{A list of objects of type \code{sccsAnalysis} as created
using the \code{\link{createSccsAnalysis}} function.}

\item{exposureOutcomeList}{A list of objects of type \code{exposureOutcome}
as created using the
\code{\link{createExposureOutcome}} function.}

\item{combineDataFetchAcrossOutcomes}{Should fetching data from the database be done one outcome at a time, or for all outcomes
 in one fetch? Combining fetches will be more efficient if there is large overlap in the subjects
that have the different outcomes.}

\item{getDbSccsDataThreads}{The number of parallel threads to use for building the
sccsData objects.}

\item{createSccsEraDataThreads}{The number of parallel threads to use for building the
sccsEraData objects.}

\item{fitSccsModelThreads}{The number of parallel threads to use for fitting the
models.}

\item{cvThreads}{The number of parallel threads to use for the cross-
validation when estimating the hyperparameter for the
outcome model. Note that the total number of CV threads
at one time could be `fitSccsModelThreads *
cvThreads`.}
}
\value{
A data frame with the following columns:
\tabular{ll}{
\verb{analysisId} \tab The unique identifier for a set of analysis choices.\cr
\verb{exposureId} \tab The ID of the target drug.\cr
\verb{outcomeId} \tab The ID of the outcome.\cr
\verb{sccsDataFolder} \tab The folder where the sccsData object is stored.\cr
\verb{sccsEraDataFolder} \tab The folder where the sccsEraData object is stored.\cr
\verb{sccsModelFile} \tab The file where the fitted SCCS model is stored.\cr
}
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses for the drug-comparator-outcomes of interest. This function will run all
specified analyses against all hypotheses of interest, meaning that the total number of outcome
models is `length(cmAnalysisList) * length(drugComparatorOutcomesList)` (if all analyses specify an
outcome model should be fitted). When you provide several analyses it will determine whether any of
the analyses have anything in common, and will take advantage of this fact. For example, if we
specify several analyses that only differ in the way the outcome model is fitted, then this
function will extract the data and fit the propensity model only once, and re-use this in all the
analysis.
}

