% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runSccsAnalyses}
\alias{runSccsAnalyses}
\title{Run a list of analyses}
\usage{
runSccsAnalyses(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  customCovariateDatabaseSchema = cdmDatabaseSchema,
  customCovariateTable = "cohort",
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "cohort",
  cdmVersion = 5,
  outputFolder = "./SccsOutput",
  sccsAnalysisList,
  exposureOutcomeList,
  combineDataFetchAcrossOutcomes = TRUE,
  getDbSccsDataThreads = 1,
  createStudyPopulationThreads = 1,
  createSccsIntervalDataThreads = 1,
  fitSccsModelThreads = 1,
  cvThreads = 1
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}}.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specify both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If \code{exposureTable = "DRUG_ERA"},
\code{exposureDatabaseSchema} is not used but assumed to be
\code{cdmDatabaseSchema}.  Requires read permissions to this database.}

\item{exposureTable}{The table name that contains the exposure cohorts.  If
\verb{exposureTable <> "DRUG_ERA"}, then expectation is \code{exposureTable}
has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
\code{outcomeTable = "CONDITION_ERA"}, \code{outcomeDatabaseSchema} is not
used.  Requires read permissions to this database.}

\item{outcomeTable}{The table name that contains the outcome cohorts.  If
outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA,
then expectation is outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{customCovariateDatabaseSchema}{The name of the database schema that is the location where
the custom covariate data is available.}

\item{customCovariateTable}{Name of the table holding the custom covariates. This table
should have the same structure as the cohort table.}

\item{nestingCohortDatabaseSchema}{The name of the database schema that is the location
where the nesting cohort is defined.}

\item{nestingCohortTable}{Name of the table holding the nesting cohort. This table
should have the same structure as the cohort table.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{outputFolder}{Name of the folder where all the outputs will written to.}

\item{sccsAnalysisList}{A list of objects of \code{sccsAnalysis} as created
using the \code{\link[=createSccsAnalysis]{createSccsAnalysis()}} function.}

\item{exposureOutcomeList}{A list of objects of type \code{exposureOutcome} as created
using the \code{\link[=createExposureOutcome]{createExposureOutcome()}} function.}

\item{combineDataFetchAcrossOutcomes}{Should fetching data from the database be done one outcome
at a time, or for all outcomes in one fetch? Combining
fetches will be more efficient if there is large overlap in
the subjects that have the different outcomes.}

\item{getDbSccsDataThreads}{The number of parallel threads to use for building the
\code{SccsData} objects.}

\item{createStudyPopulationThreads}{The number of parallel threads to use for building the
\code{studyPopulation} objects.}

\item{createSccsIntervalDataThreads}{The number of parallel threads to use for building the
\code{SccsIntervalData} objects.}

\item{fitSccsModelThreads}{The number of parallel threads to use for fitting the
models.}

\item{cvThreads}{The number of parallel threads to use for the cross-
validation when estimating the hyperparameter for the
outcome model. Note that the total number of CV threads at
one time could be \code{fitSccsModelThreads * cvThreads}.}
}
\value{
A tibble describing for each exposure-outcome-analysisId combination where the intermediary and
outcome model files can be found, relative to the \code{outputFolder}.
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses for the drug-comparator-outcomes of interest. This function will run all
specified analyses against all hypotheses of interest, meaning that the total number of outcome
models is \code{length(cmAnalysisList) * length(drugComparatorOutcomesList)} (if all analyses specify an
outcome model should be fitted). When you provide several analyses it will determine whether any of
the analyses have anything in common, and will take advantage of this fact. For example, if we
specify several analyses that only differ in the way the outcome model is fitted, then this
function will extract the data and fit the propensity model only once, and re-use this in all the
analysis.
}
