% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveExposureOutcomeList}
\alias{saveExposureOutcomeList}
\title{Save a list of exposureOutcome to file}
\usage{
saveExposureOutcomeList(exposureOutcomeList, file)
}
\arguments{
\item{exposureOutcomeList}{The exposureOutcome list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{exposureOutcome} to file. The file is in JSON format.
}

