% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveSccsAnalysisList}
\alias{saveSccsAnalysisList}
\title{Save a list of sccsAnalysis to file}
\usage{
saveSccsAnalysisList(sccsAnalysisList, file)
}
\arguments{
\item{sccsAnalysisList}{The sccsAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{sccsAnalysis} to file. The file is in JSON format.
}
