% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\name{saveSccsData}
\alias{saveSccsData}
\title{Save the cohort method data to file}
\usage{
saveSccsData(sccsData, file)
}
\arguments{
\item{sccsData}{An object of type \link{SccsData} as created using the
\link{getDbSccsData} function.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type \link{SccsData} to a file.
}
