% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{saveSccsData}
\alias{saveSccsData}
\title{Save the SCCS data to folder}
\usage{
saveSccsData(sccsData, folder)
}
\arguments{
\item{sccsData}{An object of type \code{sccsData} as generated using \code{\link{getDbSccsData}}.}

\item{folder}{The name of the folder where the data will be written. The folder should not yet
                  exist.}
}
\description{
\code{sccsData} saves an object of type sccsData to folder.
}
\details{
The data will be written to a set of files in the specified folder.
}
\examples{
# todo
}

