% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R, R/SccsEraData.R
\name{saveSccsEraData}
\alias{saveSccsEraData}
\title{Save the SCCS era data to folder}
\usage{
saveSccsEraData(SccsEraData, file)

saveSccsEraData(SccsEraData, file)
}
\arguments{
\item{SccsEraData}{An object of type \link{SccsEraData} as generated using
\code{\link[=createSccsEraData]{createSccsEraData()}}.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}

\item{sccsEraData}{An object of type \code{sccsEraData} as generated using
\code{\link{createSccsEraData}}.}

\item{folder}{The name of the folder where the data will be written. The folder should not
yet exist.}

\item{compress}{Should compression be used when saving?}
}
\value{
Returns no output.
}
\description{
\code{saveSccsEraData} saves an object of type sccsEraData to folder.

Saves an object of type \link{SccsEraData} to a file.
}
\details{
The data will be written to a set of files in the specified folder.
}
