% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversion.R
\name{saveSccsEraData}
\alias{saveSccsEraData}
\title{Save the SCCS era data to folder}
\usage{
saveSccsEraData(sccsEraData, folder, compress = FALSE)
}
\arguments{
\item{sccsEraData}{An object of type \code{sccsEraData} as generated using
\code{\link{createSccsEraData}}.}

\item{folder}{The name of the folder where the data will be written. The folder should not
yet exist.}

\item{compress}{Should compression be used when saving?}
}
\description{
\code{saveSccsEraData} saves an object of type sccsEraData to folder.
}
\details{
The data will be written to a set of files in the specified folder.
}
