% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsIntervalData.R
\name{saveSccsIntervalData}
\alias{saveSccsIntervalData}
\title{Save the cohort method data to file}
\usage{
saveSccsIntervalData(sccsIntervalData, file)
}
\arguments{
\item{sccsIntervalData}{An object of type \link{SccsIntervalData} as created using the
\link{createSccsIntervalData} function.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type \link{SccsIntervalData} to a file.
}
