% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccsData.R
\docType{class}
\name{SccsData-class}
\alias{SccsData-class}
\alias{SccsData}
\alias{show,SccsData-method}
\alias{summary,SccsData-method}
\title{SCCS Data}
\usage{
\S4method{show}{SccsData}(object)

\S4method{summary}{SccsData}(object)
}
\arguments{
\item{object}{An object of type \code{SccsData}.}
}
\description{
\code{SccsData} is an S4 class that inherits from \link[Andromeda:Andromeda]{Andromeda}. It contains information on the cases and their covariates.

A \code{SccsData} is typically created using \code{\link[=getDbSccsData]{getDbSccsData()}}, can only be saved using
\code{\link[=saveSccsData]{saveSccsData()}}, and loaded using \code{\link[=loadSccsData]{loadSccsData()}}.
}
