% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateSccsData}
\alias{simulateSccsData}
\title{Simulate SCCS data}
\usage{
simulateSccsData(nCases, settings)
}
\arguments{
\item{nCases}{The number of cases to simulate.}

\item{settings}{An object of type \code{sccsSimulationSettings} as created using the \code{
\link{createSccsSimulationSettings}}.}
}
\value{
An object of type \code{sccsData}.
}
\description{
Simulate SCCS data
}

